ESPiLight::stringToPulseTrain: correct signed/unsigned
Signed and unsigned integers are used there in a mix because the
function makes heavy use of String::indexOf which returns an int
instead uf an unsigned it. This changes the types and adds necessary
casts to have "correct" code.

+++ src/ESPiLight.h
@@ -121,7 +121,7 @@ class ESPiLight {
 
   static String pulseTrainToString(const uint16_t *pulses, size_t length);
   static int stringToPulseTrain(const String &data, uint16_t *pulses,
-                                int maxlength);
+                                size_t maxlength);
 
   static int createPulseTrain(uint16_t *pulses, const String &protocol_id,
                               const String &json);
